backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
httpx
python-dotenv
openai
boto3
requests
aiofiles
openai